Reorganized camera folder
