+++ scene/moving_camera_scene.py
@@ -3,7 +3,7 @@ from __future__ import absolute_import
 from constants import *
 
 from scene.scene import Scene
-from camera.camera import MovingCamera
+from camera.moving_camera import MovingCamera
 from topics.geometry import ScreenRectangle
 
 class MovingCameraScene(Scene):
