+++ scene/zoomed_scene.py
@@ -5,7 +5,7 @@ import numpy as np
 from scene.scene import Scene
 from animation.creation import FadeIn
 from camera.camera import Camera
-from camera.camera import MovingCamera
+from camera.moving_camera import MovingCamera
 from mobject.mobject import Mobject
 from topics.geometry import Rectangle
